Refactored WLF example plugin and added `calback_typecheck` decorator
