+++ webviz_config/generic_plugins/_example_wlf_plugin/__init__.py
@@ -1,0 +1,1 @@
+from ._plugin import ExampleWlfPlugin
