+++ webviz_config/generic_plugins/_example_wlf_plugin/_shared_view_elements/__init__.py
@@ -1,0 +1,1 @@
+from ._text_view_element import TextViewElement
