+++ webviz_config/generic_plugins/_example_wlf_plugin/views/_plot/__init__.py
@@ -1,0 +1,1 @@
+from ._view import PlotView, PlotViewElement, PlotViewSettingsGroup
