+++ webviz_config/webviz_plugin_subclasses/__init__.py
@@ -2,4 +2,4 @@ from ._settings_group_abc import SettingsGroupABC
 from ._views import ViewABC, ViewElementABC, ViewLayoutElement, LayoutElementType
 from ._layout_base_abc import LayoutBaseABC
 from ._layout_unique_id import LayoutUniqueId
-from ._webviz_callback import webviz_callback, SlotInput, EmptySlot, SlotType
+from ._callback_typecheck import callback_typecheck
